Refactoring: minor code changing
