+++ oop/index.ts
@@ -2,6 +2,6 @@ import App from "./components/App/App";
 import "./css/main.css"
 import "./css/null.css"
 
-const root = document.getElementById('root') as HTMLElement
+const root: HTMLDivElement = document.getElementById('root') as HTMLDivElement
 
 root.append(new App().render());
