+++ oop/types/JSONbinResponse.ts
@@ -1,4 +1,4 @@
-import {Item} from "./item";
+import {Item} from "./Item";
 
 export interface JSONbinResponse {
     record: {
